@@ -21,10 +21,11 @@
     "@angular/platform-browser": "^5.0.0",
     "@angular/platform-browser-dynamic": "^5.0.0",
     "@angular/router": "^5.0.0",
+    "@ngrx/store": "^4.1.1",
     "bootstrap": "^3.3.7",
     "core-js": "^2.4.1",
-    "firebase": "^4.6.0",
-    "rxjs": "^5.5.2",
+    "firebase": "4.8.0",
+    "rxjs": "^5.5.6",
     "zone.js": "^0.8.14"
   },
   "devDependencies": {
@@ -46,6 +47,6 @@
     "protractor": "~5.1.2",
     "ts-node": "~3.2.0",
     "tslint": "~5.7.0",
-    "typescript": "~2.4.2"
+    "typescript": "^2.6.2"
   }
 }
