@@ -1,12 +1,15 @@
 import { Router } from '@angular/router';
 import * as firebase from 'firebase';
 import { Injectable } from '@angular/core';
+import { LocalStorageService } from '../shared/local-storage.service';
+import { Subscription } from 'rxjs/Subscription';
+import { Unsubscribe } from 'firebase';
 
 @Injectable()
 export class AuthService {
-  token: string;
 
-  constructor(private router: Router) {}
+  constructor(private router: Router, private localStorageService: LocalStorageService) {
+  }
 
   signupUser(email: string, password: string) {
     firebase.auth().createUserWithEmailAndPassword(email, password)
@@ -20,9 +23,9 @@ export class AuthService {
       .then(
         response => {
           this.router.navigate(['/']);
-          firebase.auth().currentUser.getToken()
+          firebase.auth().currentUser.getIdToken()
             .then(
-              (token: string) => this.token = token
+              (token: string) => this.localStorageService.applyBackup({'token': token}, false, true)
             )
         }
       )
@@ -33,18 +36,22 @@ export class AuthService {
 
   logout() {
     firebase.auth().signOut();
-    this.token = null;
+    this.localStorageService.applyBackup({'token': ''});
   }
 
   getToken() {
-    firebase.auth().currentUser.getToken()
+    firebase.auth().currentUser.getIdToken()
       .then(
-        (token: string) => this.token = token
+        (token: string) => {
+          this.localStorageService.applyBackup({'token': token}, false, true);
+        }
       );
-    return this.token;
+    return this.localStorageService.getItem('token');
   }
 
   isAuthenticated() {
-    return this.token != null;
+    const token = this.localStorageService.getItem('token');
+    //this.localStorageService.consoleInfo();
+    return token != null && token.length > 0;
   }
 }
